some debugging shindigs
current state - able to distinguish if "user" property is defined in the routing context given some session (id).

ergo, we can probably just check if the "user" object is there before serving the protected endpoint request, and thus prevent an invalid session (expired / not logged in / unfamiliar) to make use of said resource.