Adding an actual Codec for "OrderToAdd"
This actually solves the issue, however instead of using the ObjectOutputStream & ObjectInputStream which can be used assuming the full class name (package included) are the same.
To avoid this issue for now, using `Json.encodeToBuffer()` as a means of encoding and decoding, having only the same classname and layout.
This may be replaced back with the original code (see attribution in file GenericCode.java files) -- once these duplicate classes are refactored out to a separate 'common' package to be comsumed in the 2 working projects.