changing User property to string
it's only purpose in this context is to identify which user name the request, and accordingly -- which file to data needs to be updated in / read from.
for this reason, the relevant username should be good enough, assuming the "User.subject()" will return the same string representation. (possible risk in case we use another auth method such as token)